nnk1

+++ OnlineShoppingSB/src/main/java/com/maven/OnlineShoppingSB/repository/OrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<OrderEntity, Integer> {
+public interface OrderRepository extends JpaRepository<OrderEntity, Long> {
     List<OrderEntity> findByUserIdAndDeletedFalse(Long userId);
     List<OrderEntity> findByDeletedFalseOrderByCreatedDateDesc();
 }
